refactor: reorganize constants and remove unused files in settings
