+++ src/app/settings/account/accountinfo/username/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import Breadcrumb from '@/components/ui/Breadcrumb';
 import { InputField } from '@/components/ui/input';
 import Button from '@/components/ui/Button';
-import { userData } from '@/constants/USER_DATA';
+import { userData } from '@/features/settings/constants/USER_DATA';
 
 export default function UsernamePage() {
   const router = useRouter();
