--- src/app/settings/account/download/page.tsx
@@ -1,11 +1,0 @@
-// This page will handle downloading user data archive
-export default function DownloadPage() {
-  return (
-    <div className="border-r border-border min-h-screen">
-      <div className="p-4">
-        <h1>Download an archive of your data</h1>
-        {/* Download archive functionality will go here */}
-      </div>
-    </div>
-  );
-}
