--- src/app/settings/security_and_account_access/two_factor/page.tsx
@@ -1,12 +1,0 @@
-export default function TwoFactorAuthPage() {
-  return (
-    <div className="p-6">
-      <h2 className="text-2xl font-bold mb-4 text-text-active">
-        Two-factor authentication
-      </h2>
-      <p className="text-text-inactive">
-        Two-factor authentication content will be added here
-      </p>
-    </div>
-  );
-}
