+++ src/features/settings/constants/SETTINGs_ITEMS.ts
@@ -28,66 +28,66 @@ export const SETTINGS_ITEMS: SettingsOption[] = [
           'See your account information like your phone number and email address.',
         path: '/settings/account/accountinfo',
       },
-      {
-        id: 'change-password',
-        label: 'Change your password',
-        description: 'Change your password at any time.',
-        path: '/settings/account/password',
-      },
-      {
-        id: 'download-archive',
-        label: 'Download an archive of your data',
-        description:
-          'Get insights into the type of information stored for your account.',
-        path: '/settings/account/download',
-      },
+      // {
+      //   id: 'download-archive',
+      //   label: 'Download an archive of your data',
+      //   description:
+      //   'Get insights into the type of information stored for your account.',
+      //   path: '/settings/account/download',
+      // },
     ],
   },
   {
-    id: 'security',
+    id: 'security_and_account_access',
     label: 'Security and account access',
     description: 'Manage your account security',
     path: '/settings/security_and_account_access',
     subOptions: [
       {
-        id: 'security-overview',
-        label: 'Security',
-        description: 'Manage your account security settings.',
-        path: '/settings/security_and_account_access',
+        id: 'change-password',
+        label: 'Change your password',
+        description: 'Change your password at any time.',
+        path: '/settings/security_and_account_access/password',
       },
-      {
-        id: 'two-factor',
-        label: 'Two-factor authentication',
-        description: 'Manage two-factor authentication options.',
-        path: '/settings/security_and_account_access/two_factor',
-      },
-      {
-        id: 'connected-apps',
-        label: 'Apps and sessions',
-        description: 'See information about when you logged into your account.',
-        path: '/settings/security_and_account_access/sessions',
-      },
+      // {
+      //   id: 'security-overview',
+      //   label: 'Security',
+      //   description: 'Manage your account security settings.',
+      //   path: '/settings/security_and_account_access',
+      // },
+      // {
+      //   id: 'two-factor',
+      //   label: 'Two-factor authentication',
+      //   description: 'Manage two-factor authentication options.',
+      //   path: '/settings/security_and_account_access/two_factor',
+      // },
+      // {
+      //   id: 'connected-apps',
+      //   label: 'Apps and sessions',
+      //   description: 'See information about when you logged into your account.',
+      //   path: '/settings/security_and_account_access/sessions',
+      // },
     ],
   },
   {
-    id: 'privacy',
+    id: 'privacy_and_safety',
     label: 'Privacy and safety',
     description: 'Manage what information you see and share',
     path: '/settings/privacy_and_safety',
     subOptions: [
-      {
-        id: 'audience',
-        label: 'Audience and tagging',
-        description:
-          'Manage what information you allow other people on X to see.',
-        path: '/settings/privacy_and_safety',
-      },
-      {
-        id: 'content-preferences',
-        label: 'Content you see',
-        description: 'Decide what you see on X based on your preferences.',
-        path: '/settings/privacy_and_safety/content',
-      },
+      // {
+      //   id: 'audience',
+      //   label: 'Audience and tagging',
+      //   description:
+      //     'Manage what information you allow other people on X to see.',
+      //   path: '/settings/privacy_and_safety',
+      // },
+      // {
+      //   id: 'content-preferences',
+      //   label: 'Content you see',
+      //   description: 'Decide what you see on X based on your preferences.',
+      //   path: '/settings/privacy_and_safety/content',
+      // },
       {
         id: 'mute-block',
         label: 'Mute and block',
@@ -95,12 +95,12 @@ export const SETTINGS_ITEMS: SettingsOption[] = [
           'Manage the accounts, words, and notifications that you want to hide.',
         path: '/settings/privacy_and_safety/mute_and_block',
       },
-      {
-        id: 'direct-messages',
-        label: 'Direct messages',
-        description: 'Manage who can message you directly.',
-        path: '/settings/privacy_and_safety/direct_messages',
-      },
+      // {
+      //   id: 'direct-messages',
+      //   label: 'Direct messages',
+      //   description: 'Manage who can message you directly.',
+      //   path: '/settings/privacy_and_safety/direct_messages',
+      // },
     ],
   },
   {
